% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_predictor.R
\name{load_predictor}
\alias{load_predictor}
\title{Load Stupid Back-off text predictor.}
\usage{
load_predictor(predtable)
}
\arguments{
\item{predtable}{a \code{sbo_predtable} object, obtained through
\code{build_predtable()}.}
}
\value{
A \code{sbo_predictor} object.
}
\description{
Load a \code{sbo_predictor} object from a \code{sbo_predtable}
}
\details{
This function restores a Stupid Back-Off text predictor
(i.e. a \code{sbo_predictor} object) from a "raw" prediction table
(a \code{sbo_predtable} object). The latter objects provide a convenient way
to store heavy N-gram models out of memory (e.g. with \code{save()}).
}
\examples{
\donttest{
# Load text predictor from sbo::twitter_predtable
p <- load_predictor(twitter_predtable)
p
predict(p, "i love")
}
}
\seealso{
\code{\link[sbo]{build_predtable}}
}
\author{
Valerio Gherardi
}
