% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_kgram_freqs.R
\name{predict.sbo_kgram_freqs}
\alias{predict.sbo_kgram_freqs}
\title{Predict method for k-gram frequency tables}
\usage{
\method{predict}{sbo_kgram_freqs}(object, input, lambda = 0.4, ...)
}
\arguments{
\item{object}{a \code{sbo_kgram_freqs} object.}

\item{input}{a length one character vector, containing the input for
next-word prediction.}

\item{lambda}{a numeric vector of length one. The back-off penalization
in Stupid Back-off algorithm.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A tibble containing the next-word probabilities for all words
in the dictionary.
}
\description{
Predictive text based on Stupid Back-off N-gram model.
}
\examples{
predict(twitter_freqs, "i love")
}
\author{
Valerio Gherardi
}
