% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_sbo_preds.R
\name{predict.sbo_preds}
\alias{predict.sbo_preds}
\title{Predict method for Stupid Back-off prediction tables}
\usage{
\method{predict}{sbo_preds}(object, newdata, L = object$L, ...)
}
\arguments{
\item{object}{a \code{sbo_preds} object.}

\item{newdata}{a character vector, containing the input for next-word prediction.}

\item{L}{a length one integer. Number of predictions for each input sentence
(maximum allowed is \code{object$L}).}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A character vector if \code{length(newdata) == 1}, otherwise a
character matrix.
}
\description{
Predictive text based on Stupid Back-off \eqn{latex}{N}-gram model.
}
\examples{
predict(twitter_sbo, "i love")
}
\author{
Valerio Gherardi
}
