% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{preprocess}
\alias{preprocess}
\title{Preprocess text corpus}
\usage{
preprocess(input, erase = "[^.?!:;'\\\\w\\\\s]", lower_case = TRUE)
}
\arguments{
\item{input}{a character vector.}
}
\value{
a character vector containing the processed output.
}
\description{
A simple text preprocessing utility.
}
\examples{
preprocess("Hi @ there! I'm using `sbo`.")
}
\author{
Valerio Gherardi
}
