% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess0}
\alias{preprocess0}
\title{Corpus preprocessing}
\usage{
preprocess0(
  text,
  to_lower = TRUE,
  replace_punct = TRUE,
  erase_extra = TRUE,
  collapse_space = TRUE,
  split_sent = FALSE,
  omit_empty = TRUE
)
}
\arguments{
\item{text}{a character vector, containing the lines of text to be processed.}

\item{to_lower}{logical. If \code{TRUE}, put everything to lower case}

\item{replace_punct}{logical. If \code{TRUE}, any punctuation including
"\code{.?!:;}" gets replaced by a single dot.}

\item{erase_extra}{logical. If \code{TRUE}, anything not alpha-numeric,
white space, . or ' gets replaced by a single white space.}

\item{collapse_space}{logical. If \code{TRUE}, multiple white spaces are
collapsed to a single white space.}

\item{split_sent}{either \code{FALSE} or a character vector of length one.
In the second case, the input gets split in sentences separated by the token
represented by \code{split_sent}.}

\item{omit_empty}{logical. Omit empty sentences resulting from splitting?
(ineffective if \code{split_sent == FALSE})}
}
\value{
A character vector. If \code{split_sent != FALSE}, each element of
the return value corresponds to a sentence, otherwise the original
structure of the input text is preserved.
}
\description{
A basic text processing function, including sentence tokenization.
}
\details{
This function provides basic utilities for preprocessing the
training corpus of a language model. The order of arguments
reflects the actual order of operations (e.g., the operations represented by
\code{replace_punct} and \code{erase_extra} do not commute).
}
\author{
Valerio Gherardi
}
