% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbo-package.R
\docType{package}
\name{sbo-package}
\alias{sbo}
\alias{sbo-package}
\title{sbo: Text Prediction via Stupid Back-Off N-Gram Models}
\description{
Utilities for building and evaluating text prediction functions based on Stupid Back-Off N-gram models.
}
\references{
The Stupid Back-Off smoothing method for N-gram models was introduced by
Brants et al., \url{https://www.aclweb.org/anthology/D07-1090/} (2007).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://vgherard.github.io/sbo/}
  \item \url{https://github.com/vgherard/sbo}
  \item Report bugs at \url{https://github.com/vgherard/sbo/issues}
}

}
\author{
Valerio Gherardi
}
\keyword{internal}
