% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{twitter_dict}
\alias{twitter_dict}
\title{Top 1000 dictionary from Twitter training set}
\format{
A character vector. Contains the 1000 most frequent words from
the example training set \code{\link[sbo]{twitter_train}}, sorted by word
rank.
}
\usage{
twitter_dict
}
\description{
Top 1000 dictionary from Twitter training set
}
\examples{
head(twitter_dict, 10)
}
\seealso{
\code{\link[sbo]{twitter_train}}
}
\keyword{datasets}
