% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{twitter_freqs}
\alias{twitter_freqs}
\title{\eqn{k}-gram frequencies from Twitter training set}
\format{
A \code{kgram_freqs} object. Contains \eqn{k}-gram frequencies from
the example training set \code{\link[sbo]{twitter_train}}.
}
\usage{
twitter_freqs
}
\description{
\eqn{k}-gram frequencies from Twitter training set
}
\seealso{
\code{\link[sbo]{twitter_train}}
}
\keyword{datasets}
