% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{twitter_predtable}
\alias{twitter_predtable}
\title{Next-word prediction tables from 3-gram model trained on Twitter training
set}
\format{
A \code{sbo_predtable} object. Contains prediction tables of a 3-gram
Stupid Back-off model trained on the example training set
\code{\link[sbo]{twitter_train}}.
}
\usage{
twitter_predtable
}
\description{
Next-word prediction tables from 3-gram model trained on Twitter training
set
}
\seealso{
\code{\link[sbo]{twitter_train}}
}
\keyword{datasets}
