% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{twitter_test}
\alias{twitter_test}
\title{Twitter test set}
\format{
A collection of 10'000 Twitter posts in English.
}
\source{
\url{https://www.kaggle.com/crmercado/tweets-blogs-news-swiftkey-dataset-4million}
}
\usage{
twitter_test
}
\description{
Twitter test set
}
\examples{
head(twitter_test)
}
\seealso{
\code{\link[sbo]{twitter_train}}
}
\keyword{datasets}
