% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{twitter_train}
\alias{twitter_train}
\title{Twitter training set}
\format{
A collection of 50'000 Twitter posts in English.
}
\source{
\url{https://www.kaggle.com/crmercado/tweets-blogs-news-swiftkey-dataset-4million}
}
\usage{
twitter_train
}
\description{
Twitter training set
}
\examples{
head(twitter_train)
}
\seealso{
\code{\link[sbo]{twitter_test}}, \code{\link[sbo]{twitter_dict}},
\code{\link[sbo]{twitter_freqs}}, \code{\link[sbo]{twitter_predtable}}
}
\keyword{datasets}
